Updated return type on ToEpocTimestamp to be long

+++ JXTPortal.Common/Extensions/DateTimeExtensions.cs
@@ -12,13 +12,13 @@ namespace JXTPortal.Common.Extensions
         /// </summary>
         /// <param name="date">The date calculate the timestamp for</param>
         /// <returns>total millisecondes between  the date and 01/01/1970</returns>
-        public static double ToEpocTimestamp(this DateTime date)
+        public static long ToEpocTimestamp(this DateTime date)
         {
             DateTime epoc = new DateTime(1970, 1, 1, 0, 0, 0);
 
             var diff = date - epoc;
 
-            return diff.TotalMilliseconds;      
+            return (long)diff.TotalMilliseconds;      
         }
 
         /// <summary>
@@ -26,7 +26,7 @@ namespace JXTPortal.Common.Extensions
         /// </summary>
         /// <param name="date">The date calculate the timestamp for</param>
         /// <returns>total millisecondes between  the date and 01/01/1970</returns>
-        public static double? ToEpocTimestamp(this DateTime? date)
+        public static long? ToEpocTimestamp(this DateTime? date)
         {
             if (!date.HasValue)
             {
@@ -37,7 +37,7 @@ namespace JXTPortal.Common.Extensions
 
             var diff = date.Value - epoc;
 
-            return diff.TotalMilliseconds;
+            return (long)diff.TotalMilliseconds;
         }
     }
 }
